pyproject.toml
[tool.poetry]
name = "hw-10"
version = "0.1.0"
description = ""
authors = ["KyryloChalov <ks7977166@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
psycopg2 = "^2.9.9"
django-braces = "^1.15.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
